visuelt og bitboard matcher

+++ Assets/Scripts/TileScripts/tileScript.cs
@@ -6,7 +6,7 @@ using UnityEngine.EventSystems;
 
 public class tileScript : MonoBehaviour
 {
-    byte DEBUG = 0;
+    int DEBUG = 0;
 
     public GameObject The_Board;
     public GameObject spawnWhitePlayer;
@@ -40,7 +40,7 @@ public class tileScript : MonoBehaviour
         if (The_Board.GetComponent<Bitboard>().bitboard[bitboardY, bitboardX] == 9)
         {
             // Get playerturn to tell which turn it is
-            byte playerturn = The_Board.GetComponent<Bitboard>().playerturn;
+            int playerturn = The_Board.GetComponent<Bitboard>().playerturn;
             MakeMove(bitboardX, bitboardY, playerturn);
         }
     }
@@ -73,7 +73,7 @@ public class tileScript : MonoBehaviour
         }
     }
 
-    void DEBUGLOG(byte[,] bitboard)
+    void DEBUGLOG(int[,] bitboard)
     {
         for (int i = 0; i < 8; i++)
         {
@@ -81,7 +81,7 @@ public class tileScript : MonoBehaviour
         }
     }
 
-    public void MakeMove (int xPos, int yPos, byte playerturn)
+    public void MakeMove (int xPos, int yPos, int playerturn)
     {
         if (DEBUG == 1)
         {
